requirements.txt
aiogram==3.6.0
replicate==0.24.0
python-dotenv==1.0.1
aiofiles==23.2.1
certifi==2024.8.30